Updating so desktop will correct misaligned frames
