+++ client/platform/desktop/backend/native/windows.ts
@@ -15,6 +15,7 @@ import {
 } from 'platform/desktop/constants';
 
 import * as viame from './viame';
+import * as mediaJobs from './mediaJobs';
 import { spawnResult } from './utils';
 
 const DefaultSettings: Settings = {
@@ -220,7 +221,7 @@ async function ffmpegCommand(settings: Settings) {
  */
 async function checkMedia(settings: Settings, file: string): Promise<CheckMediaResults> {
   await ffmpegCommand(settings);
-  return viame.checkMedia({
+  return mediaJobs.checkMedia({
     ...ViameWindowsConstants,
     setupScriptAbs: sourceString(settings),
   }, file);
@@ -230,7 +231,7 @@ async function convertMedia(settings: Settings,
   args: ConversionArgs,
   updater: DesktopJobUpdater): Promise<DesktopJob> {
   await ffmpegCommand(settings);
-  return viame.convertMedia(settings, args, updater, {
+  return mediaJobs.convertMedia(settings, args, updater, {
     ...ViameWindowsConstants,
     setupScriptAbs: sourceString(settings),
   });
