+++ client/platform/desktop/constants.ts
@@ -197,6 +197,13 @@ export interface FFProbeResults {
   }];
 }
 
+interface FFProbeFrame {
+  best_effort_timestamp_time: number;
+}
+export interface FFProbeFrameResults{
+  frames: FFProbeFrame[];
+}
+
 export type ConvertMedia =
 (settings: Settings,
   args: ConversionArgs,
